requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
sqlalchemy-pgvector==0.2.5

orjson==3.10.7
python-multipart==0.0.9

sentence-transformers==3.0.1
torch>=2.2.0

tiktoken==0.7.0
langdetect==1.0.9
regex==2024.7.24
rapidfuzz==3.9.6

python-docx==1.1.2
pymupdf==1.24.9
pytesseract==0.3.13
Pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5

spacy==3.7.5
keybert==0.8.5

openai==1.43.0

nltk==3.9.1
